tests/wpt/tests/html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.zerosource.image.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.drawImage.zerosource.image</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.zerosource.image</h1>
<p class="desc">drawImage with zero-sized source rectangle from image draws nothing without exception</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);

  function loadImage(src) {
    return new Promise((resolve, reject) => {
      const img = new Image();
      img.onload = () => resolve(img);
      img.onerror = (err) => reject(err);
      img.src = src;
    });
  }
  const img1 = await loadImage('/images/red-zerowidth.svg');
  const img2 = await loadImage('/images/red-zeroheight.svg');
  const img3 = await loadImage('/images/red-zerosize.svg');

  ctx.drawImage(img1, 0, 0, 100, 50);
  ctx.drawImage(img2, 0, 0, 100, 50);
  ctx.drawImage(img3, 0, 0, 100, 50);
  _assertPixel(canvas, 50, 25, 0, 255, 0, 255);

}, "drawImage with zero-sized source rectangle from image draws nothing without exception");
</script>
